refactor: change /deputado-federal/[name] to use nextjs13 api router
Adapt /deputado-federal/[name] page to use nextjs13 new api router. For
that, the new server component props "params" and "search params" are now used
on the page component to make it render dynamically on the server side. Also,
node-fetch is being used to make GET HTTP requests to the federal deputies
Rest API (https://dadosabertos.camara.leg.br/swagger/api.html), to
bypass what looks like to be a firewall policy that is blocking GET HTTP
requests to the API coming from the nextjs13 fetch api wrapper, probabaly due
to some additional header that is being appended to the requests, which are not
documented on the framework's API.
