+++ app/deputado-federal/[name]/_components/index.js
@@ -1,0 +1,4 @@
+import { InfoTabs } from "./InfoTabs";
+import { Header } from "./Header";
+
+export { Header, InfoTabs }
