+++ app/deputado-federal/[name]/_components/tabs/cabinet/Chart.js
@@ -1,4 +1,3 @@
-import React from "react";
 import {
   Chart as ChartJS,
   CategoryScale,
