+++ app/deputado-federal/[name]/_components/tabs/index.js
@@ -1,0 +1,15 @@
+import { ExpensesTab } from "./expenses/ExpensesTab";
+import { CabinetTab } from "./cabinet/CabinetTab";
+import { VideosTab } from "./videos/VideosTab";
+import { PersonalInfoTab } from "./personal-info/PersonalInfoTab";
+import { WorkHistoryTab } from "./work-history/WorkHistoryTab";
+import { ContactTab } from "./contact/ContactTab";
+
+export const tabs = {
+  Expenses: ExpensesTab,
+  Cabinet: CabinetTab,
+  Videos: VideosTab,
+  PersonalInfo: PersonalInfoTab,
+  WorkHistory: WorkHistoryTab,
+  Contact: ContactTab,
+};
