+++ app/deputado-federal/[name]/_components/tabs/personal-info/PersonalInfoTab.js
@@ -1,4 +1,5 @@
 import { formatDate, formatCPFCNPJ } from "@/utils";
+
 export function PersonalInfoTab({ federalDeputyBaseInfo }) {
   return (
     <>
