+++ app/deputado-federal/[name]/_components/tabs/videos/EventItem.js
@@ -1,5 +1,3 @@
-import React from "react";
-
 export default function EventItem({
   date,
   title,
