+++ app/deputado-federal/[name]/_components/tabs/work-history/ExperienceItem.js
@@ -1,5 +1,3 @@
-import React from "react";
-
 export default function ExperienceItem({
   title,
   company,
