+++ components/Footer.js
@@ -1,3 +1,5 @@
+"use client";
+
 import Link from "next/link";
 import { Tooltip, Button } from "@nextui-org/react";
