--- components/deputado-federal/[name]/index.js
@@ -1,15 +1,0 @@
-import { ExpensesTab } from "./tabs/expenses/ExpensesTab";
-import { CabinetTab } from "./tabs/cabinet/CabinetTab";
-import { VideosTab } from "./tabs/videos/VideosTab";
-import { PersonalInfoTab } from "./tabs/personal-info/PersonalInfoTab";
-import { WorkHistoryTab } from "./tabs/work-history/WorkHistoryTab";
-import { ContactTab } from "./tabs/contact/ContactTab";
-
-export const tabs = {
-  Expenses: ExpensesTab,
-  Cabinet: CabinetTab,
-  Videos: VideosTab,
-  PersonalInfo: PersonalInfoTab,
-  WorkHistory: WorkHistoryTab,
-  Contact: ContactTab,
-};
