+++ package.json
@@ -12,7 +12,6 @@
     "@heroicons/react": "^2.0.18",
     "@nextui-org/react": "^2.1.12",
     "autoprefixer": "10.4.15",
-    "axios": "^1.5.0",
     "chart.js": "^4.4.0",
     "eslint": "8.49.0",
     "eslint-config-next": "13.4.19",
@@ -20,6 +19,7 @@
     "framer-motion": "^10.16.4",
     "next": "13.4.19",
     "next-themes": "^0.2.1",
+    "node-fetch": "^3.3.2",
     "pdfjs-dist": "^3.10.111",
     "postcss": "8.4.29",
     "posthog-js": "^1.81.1",
